partial update to better layout
