+++ pkg/ditreader/constants.go
@@ -1,0 +1,30 @@
+package ditreader
+
+var nToInternal = map[string]string{
+	"uSNCreated":              "ATTq131091",
+	"uSNChanged":              "ATTq131192",
+	"name":                    "ATTm3",
+	"objectGUID":              "ATTk589826",
+	"objectSid":               "ATTr589970",
+	"userAccountControl":      "ATTj589832",
+	"primaryGroupID":          "ATTj589922",
+	"accountExpires":          "ATTq589983",
+	"logonCount":              "ATTj589993",
+	"sAMAccountName":          "ATTm590045",
+	"sAMAccountType":          "ATTj590126",
+	"lastLogonTimestamp":      "ATTq589876",
+	"userPrincipalName":       "ATTm590480",
+	"unicodePwd":              "ATTk589914",
+	"dBCSPwd":                 "ATTk589879",
+	"ntPwdHistory":            "ATTk589918",
+	"lmPwdHistory":            "ATTk589984",
+	"pekList":                 "ATTk590689",
+	"supplementalCredentials": "ATTk589949",
+	"pwdLastSet":              "ATTq589920",
+}
+
+var accTypes = map[int32]string{
+	0x30000000: "SAM_NORMAL_USER_ACCOUNT",
+	0x30000001: "SAM_MACHINE_ACCOUNT",
+	0x30000002: "SAM_TRUST_ACCOUNT",
+}
