+++ pkg/esent/structures.go
