+++ pkg/winregistry/winregistry.go
